feat(frontend): i18n, accessibility, bundle, and RSC upgrades
Closes #1187 #1188 #1189 #1190

#1187 — Internationalization (i18n) for Notification Center
- Added 'notificationCenter' namespace to messages/en.json, es.json, pt.json
- Replaced all hardcoded strings in NotificationCenter.tsx with useTranslations()
- Keys: buttonLabel, buttonLabelWithCount, panelLabel, heading, unreadCount,
  noAlerts, alertLabel, dismissLabel, notificationLabel, timestampLabel

#1188 — Refactor Notification Center (accessibility + responsive + clean CSS)
- Extracted NotificationPanel as a separate exportable component
- Added WCAG-compliant ARIA: aria-haspopup='dialog', aria-expanded, aria-modal,
  aria-live/aria-atomic on unread count, role=list/listitem
- Min 44x44px touch target on trigger button (WCAG 2.5.5)
- Dismiss button always visible on keyboard focus (opacity-0 group-hover/focus)
- Responsive panel: full-width on mobile, capped at max-w-xs sm:w-80
- Clean Tailwind classes replacing all inline/arbitrary styles where possible
- focus-visible ring for keyboard navigation

#1189 — Optimize client-side bundle size for Notification Center
- Changed layout.tsx to use next/dynamic with ssr:false for NotificationCenter
- Defers the component JS until after initial hydration
- Exported NotificationPanel separately to allow further code-splitting

#1190 — Migrate User Profile Settings Widget to React Server Components
- Created settings/SettingsWidget.tsx ('use client') with all interactive logic
- Replaced settings/page.tsx with a React Server Component wrapper that:
  - Has no 'use client' directive (runs on server / edge)
  - Exports page metadata (title, description) via Next.js Metadata API
  - Renders SettingsWidget as a client component child

Misc:
- Updated NotificationCenter tests to mock next-intl and align assertions
  with i18n key-based output
- Updated settings/page.test.tsx to import from SettingsWidget directly
- Updated .gitignore and frontend/.gitignore to exclude test-results/,
  playwright snapshots, tsconfig.tsbuildinfo, and output.txt files
