+++ frontend/.gitignore
@@ -1,2 +1,21 @@
 node_modules/
 package-lock.json
+
+# Next.js build output
+.next/
+out/
+
+# TypeScript incremental build cache
+tsconfig.tsbuildinfo
+
+# Playwright test artifacts
+test-results/
+playwright-report/
+
+# Visual regression snapshots (generated at runtime, not committed)
+tests/e2e/*.spec.ts-snapshots/
+tests/e2e/*.spec.tsx-snapshots/
+*-snapshots/
+
+# Misc output files generated by scripts/debug runs
+output.txt
